django/app/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
alembic==0.9.3.dev0
amqp==2.5.2
apturl==0.5.2
arrow==0.15.2
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
Babel==2.4.0
beautifulsoup4==4.8.1
billiard==3.6.3.0
binaryornot==0.4.4
blinker==1.4
Brlapi==0.6.6
bs4==0.0.1
cchardet==2.1.6
celery==4.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
configparser==5.0.0
constantly==15.1.0
cookiecutter==1.6.0
cookiejar==0.0.2
cryptography==3.0
cssselect==1.1.0
cupshelpers==1.0
defer==1.0.6
demjson==2.2.4
distro-info===0.18ubuntu0.18.04.1
Django==3.0.8
django-celery-beat==1.5.0
django-filter==2.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.1
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.7.0
incremental==17.5.0
isort==4.3.20
itsdangerous==0.24
Jinja2==2.10.3
jinja2-time==0.2.0
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
ldap3==2.4.1
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
minio==5.0.7
multidict==4.7.6
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pager==3.3
pandas==1.0.3
paramiko==2.0.0
parsel==1.6.0
passlib==1.7.1
pexpect==4.2.1
Pillow==5.1.0
poyo==0.5.0
pproxy==2.3.4
protobuf==3.0.0
psutil==5.6.7
psycopg2==2.7.7
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==2.0.2
pyinotify==0.9.6
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-crontab==2.4.1
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
raven==6.10.0
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
Scrapy==1.6.0
scrapy-crawlera==1.2.2
scrapy-splash==0.7.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.15.0
soupsieve==1.9.4
speaklater==1.4
SQLAlchemy==1.1.11
sqlparse==0.3.1
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
Twisted==20.3.0
typed-ast==1.3.5
typing==3.7.4.1
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.6
usb-creator==0.3.3
vboxapi==1.0
vine==1.3.0
virtualenv==15.1.0
w3lib==1.22.0
wadllib==1.3.2
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.11.1
WTForms==2.1
xkit==0.0.0
xxhash==1.4.4
yarl==1.4.2
zipp==3.1.0
zope.interface==5.1.0
